+++ src/main/java/org/decimal4j/arithmetic/FloatConversion.java
@@ -1,7 +1,7 @@
 /**
  * The MIT License (MIT)
  *
- * Copyright (c) 2015 decimal4j (tools4j), Marco Terzer
+ * Copyright (c) 2016 decimal4j (tools4j), Marco Terzer
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -121,13 +121,13 @@ final class FloatConversion {
 			if (x >= 0.0f || isMathematicalInteger(x)) {
 				return x;
 			} else {
-				return x - 1.0f;
+				return (long)x - 1.0f;
 			}
 		case CEILING:
 			if (x <= 0.0f || isMathematicalInteger(x)) {
 				return x;
 			} else {
-				return x + 1.0f;
+				return (long)x + 1.0f;
 			}
 		case DOWN:
 			return x;
@@ -135,7 +135,7 @@ final class FloatConversion {
 			if (isMathematicalInteger(x)) {
 				return x;
 			} else {
-				return x + Math.copySign(1.0f, x);
+				return (long)x + Math.copySign(1.0f, x);
 			}
 		case HALF_EVEN:
 			return rint(x);
